Updates install step in CI
