+++ README.md
@@ -23,8 +23,7 @@ npm start
 
 Alternatively, you may run individual steps within `npm start` one at a time:
 ```bash
-npm install # Installs Node & web dependencies
-npm run install-db-data # Download data for database
+npm install # Installs Node, web dependencies, & raw data for database
 npm run init-db # Creates & populates database tables
 npm run app # Compiles Typescript, bundles application, then launches Electron app
 ```
