+++ package.json
@@ -7,11 +7,10 @@
     "license": "GNU General Public License v3.0",
     "type": "module",
     "scripts": {
-        "install-bootstrap": "npx tsx ci/install-bootstrap.ts",
+        "install-resources-data": "npx tsx ci/install.ts",
         "rebuild-sqlite3": "electron-rebuild -f -w sqlite3",
-        "postinstall": "npm run install-bootstrap && npm run rebuild-sqlite3",
-        "install-db-data": "node --disable-warning=MODULE_TYPELESS_PACKAGE_JSON db/install-db-data.js",
-        "init-db": "node --disable-warning=MODULE_TYPELESS_PACKAGE_JSON db/init-db.js",
+        "postinstall": "npm run install-resources-data && npm run rebuild-sqlite3",
+        "init-db": "npx tsx ci/init-db.ts",
         "prep-build-dir": "rm -rf build && mkdir build && cp -r installed-resources/* build && cp -r custom-resources/* build",
         "compile-typescript": "tsc",
         "compile-react": "webpack --mode development --config webpack.react.config.js",
